+++ examples/customize/build_graph/components/splitters/fixed_size_splitter.py
@@ -6,9 +6,10 @@ from neo4j_graphrag.experimental.components.types import TextChunks
 
 async def main() -> TextChunks:
     splitter = FixedSizeSplitter(
-        # optionally, configure chunk_size and chunk_overlap
+        # optionally, configure chunk_size, chunk_overlap, and approximate flag
         # chunk_size=4000,
         # chunk_overlap=200,
+        # approximate = False
     )
     chunks = await splitter.run(text="text to split")
     return chunks